Implemented running of original boot image (+ utils).
Implemented loading and running of original boot image. Original image is moved back to its correct place. It will also restore the metadata (creation/modification time) back to the original boot image, as it should have been moved in the previous boot. Also added capability of unloading image (to be used if image can't be started).