Hvloader + hv's vmexit hook.
Hvloader + hv's vmexit hook. Added hvloader hook (applied in winload's ImgpLoadPEImage hook when hvloader is being loaded). The routine which is hooked in hvloader is called in the hypervisor launch routine, and it is responsible for jumping to a gadget (3rd param - which jmps to hypervisor image's entry point after loading hyperv's cr3), which allows us to tweak the remapped hyper-v image and also the page tables that it will execute under. We patch hyper-v's page tables with an (executable) identity map, so our driver will be mapped later on (and we can also use it for easy host physical memory access down the line). We also patch a relative call to the vmexit handler to a code cave (in which a long jump is done to our driver), hence, giving us execution in the vmexit handler, although calling the original routine crashes at the moment.

Unrelated changes: added page allocation helper (needed for 4kb aligned allocations for page table structures), and made hook data embedded within the uefi-boot binary itself, instead of a runtime pool allocation.