Fixed vmexit original handler call crash.
Fixed vmexit original handler call crash. Resolving of original handler's address had a logic error, now we apply the signed value of the rva instead of the unsigned value. Added restoration of original rva (to the relative call of the vmexit handler) on first call, as we won't need the hook later (will load a seperate module for this).