Added Hyper-V attachment loading.
Added Hyper-V attachment loading (I have basically created a simple image loader which copies the sections to the runtime allocation and relocates the image). The 'Hyper-V attachment' is a separate image in the project, residing in a SLAB allocation, which will be controlling Hyper-V (at the moment, only through its vmexit hook). This module is loaded from disk then deleted in the uefi-boot image entry before passing execution to the original bootmgfw image. At the moment, the hyper-v attachment module is residing in the identity map in the hyper-v cr3, it doesn't have independent PTEs allocated for itself. Instead of the vmexit handler hook pointing to our uefi-boot image, it will now point to the hyper-v attachment image. I have also refactored the SLAB allocation helper to have a wrapper which translates the returned virtual address pointing to the SLAB allocation to its physical one, and we do this through the use of winload's BlMmTranslateVirtualAddressEx.