Simple cpuid call handler + usermode implementation.
Simple cpuid call handler + usermode implementation. The usermode app will cpuid with rcx = 1337, if this is found by the vmexit handler detour (as in: cpuid and rcx = 1337) then itll advance the rip and set the guest trap frame's rax to 1337 without not call the original vmexit handler (as this isn't a vmexit hyper-v needs to process). If the vmexit doesn't fit our requirements, then the original vmexit handler will be invoked. At the moment, this will only work for Intel CPUs.