Added guest phys -> host phys translation.
Added guest physical address -> host physical address translation. In the root partition, guest and host physical addresses are 1:1, unless hidden by the hypervisor by slat. This is achieved by walking the EPT paging structures to translate the guest physical address to the host physical address. Currently only implemented for Intel machines.

I also changed the identity map in the hyper-v cr3 to be in the 255th pml4e (to avoid hyper-v's usage of the 0 index pml4e + is more elegant for them not to be 1:1), so now the hyper-v attachment is also relocated to fit that address space.