Added heap manager.
Added heap manager. This resides in the hyper-v attachment. The initial physical buffer for the heap is passed into the entry point of the hyper-v attachment, where it is then set up. We can now allocate and free memory from this heap as we execute. There is also a mutex implemented for when the manager is trying to allocate pages, this is so that two cores don't try take the same page at once.