Added SLAT code hooks + memory manager updates.
Added SLAT code hooks + memory manager updates. When a slat violation is raised, we check if it is a part of a physical page we are hooking, and if it is, then (Intel specific): we set the page to rw- (not executable, but readable and writable) protection and to PFN to the original (so when read/write operation occurs, it happens to the original host page) if it was caused by a read / write operation, and if it was an instruction fetch (execution), then we set the page to --x (not readable or writable, but executable) and the PFN to the shadow host page. This will allow us to control what host page a guest page executes from, whilst still retaining legitimacy from the eyes of the guest, as whenever a read or write operation occurs. itll lead to the original host page. The usermode app can hypercall to set up SLAT code hooks.

Added hypercall which allows the caller to get their current cr3 value.

Heap manager was changed to a list-based allocator fixed block size of 0x1000, so we can ensure page alignment. This is also a lot faster implementation than we had before, and we also don't allocate anything at the moment that isnt 0x1000 in size, so  wastage of part of a block is not an issue.