More commands + address aliases.
More commands + address aliases. Addresses such as the nt kernel base address or the current process cr3 are able to be accessed easily by aliasing. When a user inputs say a translation and wants to use current process cr3, in place of the cr3 value, they would put 'current_cr3'. I also had to implementing process module enumeration (in this case, the system process) (raising of privileges is needed in 24h2 to get the base address of ntoskrnl). The commands I have added is the read, write, and copy of virtual memory.