Added ability to hook kernel routines.
Added ability to hook kernel routines. This is done by a mixture of slat and inline hooks. On initialization, we find a page in ntoskrnl (pe headers parsing done using portable-executable library by papstuc, so credit there) which we can ept hook and use as our 'gadget page' (where detours containing retpolines and the assembly that users want to execute reside). When we want to hook a kernel routine, we ept hook the page, and set a jump to a corresponding gadget in the ntoskrnl gadget page. We have to execute original bytes, so we use zydis to decode the next few instructions and find the aligned length, and this then gets executed in the detour. Flow is then returned back to the original routine. I have also implemented the command interface for this (including removal of aforementioned hooks).

I had to fix a logic error in the removal of ept hooks where it would essentially erase part of the linked list containing in-use hooks.