Added guest state logging (used for hooks)
Added guest state logging (used for hooks). When a user wants to place a kernel hook, they can optionally add the monitor flag, which will log the trap frame of the routine that is being hooked on entry, these logs are done by vmexiting and the hyperv-attachment captures them. The logs are then flushed by the usermode program, so they can be processed and displayed.