Added boot services AllocatePages hook
Added boot services AllocatePages hook. This is only applied before we allocate slab pages through winload.efi, as the pages will be specified as loader memory type, meaning they can be discarded later. In our hook, we set the memory type to runtime so that our allocations will not be repurposed. The hook is removed right after the allocation takes place.