Firing NMIs for SLAT cache invalidation + more
When we add or remove a slat hook, we will need to invalidate the SLAT caches (eg. due to splitting of 2mb SLAT entry to 4kb entries).

This invalidation must be on all virtual host processors, so we fire non maskable interrupts via an xapic / x2apic interface to all host virtual processors.

If a host processor is in guest state, then a vmexit will occur, where we will then process the exit and do the invalidation.

If a host processor is NOT in guest state, then the NMI will be delivered by the host's interrupt descriptor table (IDT), where we have added a handler at vector 2 so that we can then invalidate the caches before returning execution back to the interrupted context.

More - added a primary and secondary vmexit key (all fit within eax), so that an accidental command process is unlikely. I also changed the ntoskrnl section we use for the hook detour holder to one which isn't pageable.