Hook improvements + log improvements.
Log improvements - we now read a set amount of 64 bit numbers from the guest rsp and insert it into the logs so it can be outputted later.

Hook improvements - we now allow a set assembled bytes to be specified to be executed after the original bytes are called. Furthermore, I fixed some bugs with rip relative instructions being displaced by the increased length of operands. In addition, I fixed memory related rip relative operands (in eg. mov, cmp) having a logic error where the address had to be dereferenced (we now find a register not used in any operands in that instruction to fix and push / pop it and use it to load the new absolute value to operate on).