New command + bug fixes
Added command to dump a kernel module from memory to disk at a specified path in the usermode app.

I also fixed a bug in the hypercall for memory operations where the destination would be calculated wrong and the process would still proceed, it now is terminated if the destination is calculated wrongly.

I also made it so when translating guest physical addresses, SLAT hooked pages return the original page frame number, not our shadow one. This is on Intel only as we have 2 slat cr3s on AMD.