Made Hyper-V hooks more robust.
Made Hyper-V hooks more robust. Firstly, we find the true beginning/end of the .text section by walking the Hyper-V cr3 and finding the first and last executable page around Hyper-V's entry point virtual address. We then use these boundaries to signature scan, which is safer than blindly assuming where the .text section ends or begins like we did before.

In addition, we now check for the AMD signature first, as it is more uncommon, and will be far less likely to be found in an Intel Hyper-V binary.